fix: change in Environment by creating EnvironmentProvider
