+++ packages/client/src/home/environmentSelector.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import MenuItem from '@mui/material/MenuItem';
 import FormControl from '@mui/material/FormControl';
 import Select, { SelectChangeEvent } from '@mui/material/Select';
-import { useEnvironment } from '../hooks';
+import { useEnvironment } from '../context/EnvironmentProvider';
 
 export default function EnvironmentSelector({ environmentList }) {
     const { environment, setEnvironment } = useEnvironment();
